requirements.txt
# BoarderframeOS Core Dependencies

# Web Framework and API
fastapi==0.104.1
flask==3.0.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2

# Database
asyncpg==0.29.0
pgvector==0.2.4
psycopg2-binary==2.9.9

# Redis and Caching
redis==5.0.1
aioredis==2.0.0

# Agent Framework - Updated for Brain + LangGraph Integration
langgraph==0.2.40
langchain==0.3.10
langchain-openai==0.2.10
langchain-anthropic==0.2.8
langchain-community==0.3.8
langchain-core==0.3.25
crewai==0.1.0

# Extended LLM Provider Support
langchain-google-genai==2.0.0
langchain-cohere==0.3.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0

# The Brain Dependencies
litellm==1.50.29
qdrant-client==1.12.1

# Enhanced Monitoring
langsmith==0.1.147
agentops==0.3.25

# Additional Redis for Streams
redis[hiredis]==5.2.0

# Vector and ML
numpy==1.24.3
openai==1.3.8
anthropic==0.7.8

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
jsonschema==4.19.2
click==8.1.7
rich==13.7.0
typer==0.9.0
psutil==5.9.6

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Background Tasks
celery==5.3.4

# File Processing
pillow==10.1.0
python-magic==0.4.27

# SQLite compatibility (for migration)
aiosqlite==0.19.0
